+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 * Enhancement: Pass request to CustomSamplingContext in Spring integration (#1172)
 * Fix: Free Local Refs manually due to Android local ref. count limits
 * Fix: Bring back support for setting transaction name without ongoing transaction (#1183)
+* Enhancement: Add overload for `transaction/span.finish(SpanStatus)` (#1182)
+
 # 4.0.0-alpha.3
 
 * Feat: Add maxAttachmentSize to SentryOptions (#1138)